move rio commands to individual files
