+++ rasterio/rio/env.py
@@ -1,0 +1,26 @@
+"""Fetch and edit raster dataset metadata from the command line."""
+
+import logging
+
+import click
+
+import rasterio
+import rasterio.crs
+
+
+@click.command(short_help="Print information about the rio environment.")
+@click.option('--formats', 'key', flag_value='formats', default=True,
+              help="Enumerate the available formats.")
+@click.pass_context
+def env(ctx, key):
+    """Print information about the Rasterio environment: available
+    formats, etc.
+    """
+    verbosity = (ctx.obj and ctx.obj.get('verbosity')) or 1
+    logger = logging.getLogger('rio')
+    stdout = click.get_text_stream('stdout')
+    with rasterio.drivers(CPL_DEBUG=(verbosity > 2)) as env:
+        if key == 'formats':
+            for k, v in sorted(env.drivers().items()):
+                stdout.write("%s: %s\n" % (k, v))
+            stdout.write('\n')
