+++ rasterio/rio/stack.py
