+++ tests/test_rio_stack.py
@@ -1,15 +1,14 @@
-import click
 from click.testing import CliRunner
 
 import rasterio
-from rasterio.rio import bands
+from rasterio.rio.stack import stack
 
 
 def test_stack(tmpdir):
     outputname = str(tmpdir.join('stacked.tif'))
     runner = CliRunner()
     result = runner.invoke(
-        bands.stack,
+        stack,
         ['tests/data/RGB.byte.tif', outputname],
         catch_exceptions=False)
     assert result.exit_code == 0
@@ -21,7 +20,7 @@ def test_stack_list(tmpdir):
     outputname = str(tmpdir.join('stacked.tif'))
     runner = CliRunner()
     result = runner.invoke(
-        bands.stack,
+        stack,
         ['tests/data/RGB.byte.tif', '--bidx', '1,2,3', outputname])
     assert result.exit_code == 0
     with rasterio.open(outputname) as out:
@@ -32,7 +31,7 @@ def test_stack_slice(tmpdir):
     outputname = str(tmpdir.join('stacked.tif'))
     runner = CliRunner()
     result = runner.invoke(
-        bands.stack, 
+        stack,
         [
             'tests/data/RGB.byte.tif', '--bidx', '..2',
             'tests/data/RGB.byte.tif', '--bidx', '3..',
@@ -46,7 +45,7 @@ def test_stack_single_slice(tmpdir):
     outputname = str(tmpdir.join('stacked.tif'))
     runner = CliRunner()
     result = runner.invoke(
-        bands.stack, 
+        stack,
         [
             'tests/data/RGB.byte.tif', '--bidx', '1',
             'tests/data/RGB.byte.tif', '--bidx', '2..',
@@ -61,7 +60,7 @@ def test_format_jpeg(tmpdir):
     outputname = str(tmpdir.join('stacked.jpg'))
     runner = CliRunner()
     result = runner.invoke(
-        bands.stack,
+        stack,
         ['tests/data/RGB.byte.tif', outputname, '--format', 'JPEG'])
     assert result.exit_code == 0
 
@@ -70,6 +69,6 @@ def test_error(tmpdir):
     outputname = str(tmpdir.join('stacked.tif'))
     runner = CliRunner()
     result = runner.invoke(
-        bands.stack,
+        stack,
         ['tests/data/RGB.byte.tif', outputname, '--driver', 'BOGUS'])
     assert result.exit_code == 1
